% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{calc_mean_gcnt}
\alias{calc_mean_gcnt}
\title{Calcula o Valor Esperado para a Distribui\enc{çã}{ca}o Gamma Count}
\usage{
calc_mean_gcnt(lambda, alpha, tol = 1e-05)
}
\arguments{
\item{lambda}{Parâmetro de locação da distribuição Gamma Count,
\eqn{\lambda > 0}. Quando \eqn{\alpha = 1}, o parâmetro
\eqn{\lambda = E(Y)} é a média.}

\item{alpha}{Parâmetro de dispersão da distribuição Gamma Count,
\eqn{\alpha > 0}.}

\item{tol}{Tolerância para interromper a procura pelo valor de
\code{ymax}, valor cuja probabilidade correspondente é inferior a
\code{tol}, para valores os valores de \code{lambda} e
\code{alpha} informados.}
}
\value{
Um vetor de tamanho igual ao do maior vetor, \code{lambda} ou
    \code{alpha} com os valores correspondentes de \eqn{\mu}.
}
\description{
Função para calcular a média do tipo \eqn{E(Y) = \mu =
    \sum y\cdot \Pr(y)} para uma variável aleatória Gamma Count a
    partir dos parâmetros \eqn{\lambda > 0} e \eqn{\alpha > 0}.
}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}.
}

