% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp.R
\name{calc_var_cmp}
\alias{calc_var_cmp}
\title{Calcula o Valor da Variância para a Distribuição
    Conway-Maxwell-Poisson}
\usage{
calc_var_cmp(lambda, nu, sumto, tol = 1e-05)
}
\arguments{
\item{lambda}{Valor do parâmetro \eqn{\lambda} da distribuição
COM-Poisson. Quando \eqn{\nu = 1}, o parâmetro \eqn{\lambda =
E(Y)} é a média.}

\item{nu}{Valor do parâmetro \eqn{\nu} da distribuição COM-Poisson.}

\item{sumto}{Número de incrementos a serem considerados para a
cálculo da constante normalizadora Z.}

\item{tol}{Tolerância para interromper a procura pelo valor de
\code{ymax}, valor cuja probabilidade correspondente é inferior a
\code{tol}, para valores os valores de \code{lambda} e \code{nu}
informados.}
}
\value{
Um vetor de tamanho igual ao do maior vetor, \code{lambda} ou
    \code{nu} com os valores correspondentes de \eqn{\mu}.
}
\description{
Função para calcular a variância do tipo \eqn{V(Y) =
    E(Y^2) - E^2(Y) = \sum y^2\cdot \Pr(y) - \left ( \sum y\cdot
    \Pr(y) \right )^2} para uma variável aleatória COM-Poisson a
    partir dos parâmetros \eqn{\lambda > 0} e \eqn{\nu \geq 0}.
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}

