% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp.R
\name{cmp}
\alias{cmp}
\title{Ajuste do Modelo Conway-Maxwell-Poisson}
\usage{
cmp(formula, data, start = NULL, sumto = NULL, ...)
}
\arguments{
\item{formula}{Um objeto da classe \code{\link{formula}}. Se
necessária a inclusão de \emph{offset} deve-se indicá-lo como
\code{\link{offset}}.}

\item{data}{Um objeto de classe \code{data.frame}, cujo contém as
variáveis descritas na \code{formula}.}

\item{start}{Um vetor de parâmetros do modelo tomados como valores
iniciais para o algoritmo iterativo. Se \code{NULL} os parâmetros
de um modelo log-linear Poisson e \eqn{\phi = 0} são tomados como
valores iniciais para o preditor e para o parâmetro de dispersão.}

\item{sumto}{Número de incrementos a serem considerados para a soma
da constante normalizadora. Como padrão, \code{NULL} o número de
incrementos é o valor inteiro de \eqn{(\max y)^1.2}.}

\item{...}{Argumentos opcionais do framework de maximização numérica
\code{\link[bbmle]{mle2}}.}
}
\value{
Um objeto de classe \code{mle2}, retornado da função de
    \code{\link[bbmle]{mle2}}, usada para ajuste de modelos por
    máxima verossimilhança.
}
\description{
Estima os parâmetros de um modelo COM-Poisson sob a
    otimização da função de log-verossimilhança. A sintaxe
    assemelha-se com a função \code{\link{glm}} (Generalized Linear
    Models).
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}

