% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp.R
\name{convergencez}
\alias{convergencez}
\title{Avaliação da Convergência da Constante Normalizadora}
\usage{
convergencez(model, tol = 1e-04, incremento = 10, maxit = 50,
  plot = TRUE)
}
\arguments{
\item{model}{Objeto resultante da função \code{\link[MRDCr]{cmp}}.}

\item{tol}{Critério de parada do algoritmo, representa o valor
tolerado para a diferença do valor de \eqn{Z(\lambda, \phi)}
entre duas iterações. O valor padrão é 1e-4}

\item{incremento}{Número de incrementos da soma a serem considerados
a cada iteração. Padrão definido como 10, ou seja, a cada
iteração 10 passos incrementos são somados a Z.}

\item{maxit}{Número máximo de iterações a serem realizadas pelo
algoritmo. Se este número for atingido e o critério de tolerância
não for atendido, uma mensagem de aviso será exibida.}

\item{plot}{Argumento lógico. Se \code{TRUE} (padrão) os gráficos dos
incrementos da constante são exibidos.}
}
\value{
Uma lista com os incrementos das constantes Z,
    \eqn{Z(\lambda, \phi)} da distribuição COM-Poisson, calculados
    para cada observação.
}
\description{
Avalia a convergência da constante de normalização de
    um modelo COM-Poisson definida por: \deqn{Z = \sum
    \frac{\lambda^i}{(i!)^\nu}}, em que o parâmetro \eqn{\nu} é
    tomado como \eqn{\exp{\phi}}.
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}

