% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp.R
\name{dcmp}
\alias{dcmp}
\title{Probabilidades do Modelo Conway-Maxwell-Poisson}
\usage{
dcmp(y, lambda, nu, sumto)
}
\arguments{
\item{y}{Valor da variável de contagem.}

\item{lambda}{Valor do parâmetro \eqn{\lambda} da distribuição
COM-Poisson.}

\item{nu}{Valor do parâmetro \eqn{\nu} da distribuição COM-Poisson.}

\item{sumto}{Número de incrementos a serem considerados para a
cálculo da constante normalizadora Z.}
}
\description{
Calcula as probabilidades para uma variável aleatória
    distribuída conforme modelo COM-Poisson.

\deqn{p(y,\lambda,\nu) =
    \frac{\lambda^y}{(y!)^\nu Z(\lambda, \nu)}
}

em que \eqn{Z(\lambda, \nu)} é a constante de normalização definida
    por \eqn{\sum_{j=0}^{\infty} \frac{\lambda^j}{(j!)^\nu}}.  Nesta
    implementação o número de incrementos considerados para cálculo
    dessa constante é definido por \code{sumto}. \eqn{\lambda > 0} e
    \eqn{\nu \geq 0} são os parâmetros da distribuição.
}
\examples{
dpois(5, lambda = 5)
dcmp(5, lambda = 5, nu = 1, sumto = 20)

probs <- data.frame(y = 0:30)
within(probs, {
    py0 <- dpois(y, lambda = 15)
    py1 <- dcmp(y, lambda = 15, nu = 1, sumto = 50)
    py2 <- dcmp(y, lambda = 915, nu = 2.5, sumto = 50)
    py3 <- dcmp(y, lambda = 2.2, nu = 0.3, sumto = 50)
    plot(py0 ~ y, type = "h",
         ylim = c(0, max(c(py0, py2, py3))),
         ylab = expression(Pr(Y == y)))
    points(y + 0.1, py1, type = "h", col = 2)
    points(y - 0.3, py2, type = "h", col = 3)
    points(y + 0.3, py3, type = "h", col = 4)
    legend("topleft", bty = "n",
           col = c(1:4), lty = 1,
           legend = expression(
               Poisson(lambda == 15),
               CMP(lambda == 15, nu == 1),
               CMP(lambda == 915, nu == 2.5),
               CMP(lambda == 2.2, nu == 0.3)))
})
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}
