% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcnt.R
\name{dgcnt}
\alias{dgcnt}
\title{Probabilidades do Modelo Gamma Count}
\usage{
dgcnt(y, lambda, alpha)
}
\arguments{
\item{y}{Valor da variável de contagem.}

\item{lambda}{Valor do parâmetro \eqn{\lambda} que é a média da
distribuição do tempo entre eventos.}

\item{alpha}{Valor do parâmetro \eqn{\alpha} que é o parâmetro de
dispersão.}
}
\value{
Retorna uma probabilidade, ou seja \eqn{\Pr(Y = y | \lambda,
    \alpha) = p(y, \lambda, \alpha)}.
}
\description{
Calcula as probabilidades para uma variável aleatória
    com distribuição Gamma Count:

\deqn{p(y,\lambda,\alpha) =
  \left(\int_{0}^{1}
  \frac{(\alpha\lambda)^{y\alpha}}{\Gamma(y\alpha)}\,
  u^{y\alpha-1}
  \exp\{-\alpha\lambda u\}\, \textrm{d}u \right)
  - \left(\int_{0}^{1}
  \frac{(\alpha\lambda)^{y\alpha}}{\Gamma((y+1)\alpha)}\,
  u^{(y+1)\alpha-1}
  \exp\{-\alpha\lambda u\}\, \textrm{d}u \right),}

em que \eqn{\lambda > 0} é a média da variável aleatória tempo entre
    eventos e \eqn{\alpha > 0} é o parâmetro de dispersão.
}
\examples{

dpois(5, lambda = 5)
dgcnt(5, lambda = 5, alpha = 1)

probs <- data.frame(y = 0:30)
within(probs, {
    py0 <- dpois(y, lambda = 15)
    py1 <- dgcnt(y, lambda = 15, alpha = 1)
    py2 <- dgcnt(y, lambda = 15, alpha = 0.5)
    py3 <- dgcnt(y, lambda = 15, alpha = 1.5)
    plot(py0 ~ y, type = "h",
         ylim = c(0, max(c(py0, py2, py3))),
         ylab = expression(Pr(Y == y)))
    points(y + 0.1, py1, type = "h", col = 2)
    points(y - 0.3, py2, type = "h", col = 3)
    points(y + 0.3, py3, type = "h", col = 4)
    legend("topleft", bty = "n",
           col = c(1:4), lty = 1,
           legend = expression(
               Poisson(lambda == 15),
               GC(lambda == 15, alpha == 1),
               GC(lambda == 15, alpha == 0.5),
               GC(lambda == 15, alpha == 1.5)))
})

}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}.
}

