% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgnz.R
\name{dpgnz}
\alias{dpgnz}
\title{Probabilidades do Modelo Poisson Generalizado}
\usage{
dpgnz(y, lambda, alpha)
}
\arguments{
\item{y}{Valor da variável de contagem.}

\item{lambda}{Valor do parâmetro \eqn{\lambda} que é a média da
distribuição .}

\item{alpha}{Valor do parâmetro \eqn{\alpha} que é o parâmetro de
dispersão.}
}
\value{
Retorna uma probabilidade, ou seja \eqn{\Pr(Y = y | \lambda,
    \alpha) = p(y, \lambda, \alpha)}.
}
\description{
Calcula as probabilidades para uma variável aleatória
    com distribuição Poisson Generalizada na parametrização
    \eqn{\lambda-\alpha}:

\deqn{p(y,\lambda,\alpha) =
  \left(\frac{\lambda}{1+\alpha\lambda}\right)^{y}
  \,\frac{(1+\alpha y)^{y-1}}{y!}
  \exp\left\{-\lambda\left(
  \frac{1+\alpha y}{1+\alpha\lambda}\right)\right\},
}

em que \eqn{\lambda > 0} é a média da variável aleatória e
    \eqn{\alpha} é o parâmetro de dispersão, sendo que \eqn{V(Y) =
    \lambda (1 + \alpha\lambda)^2}. O espaço paramétrico de
    \eqn{\alpha} depende de \eqn{\lambda} e \eqn{y} pois
    \eqn{1+\alpha\lambda > 0} e \eqn{1+\alpha y > 0}.
}
\examples{

dpois(5, lambda = 5)
dpgnz(5, lambda = 5, alpha = 0)

probs <- data.frame(y = 0:30)
within(probs, {
    py0 <- dpois(y, lambda = 15)
    py1 <- dpgnz(y, lambda = 15, alpha = 0)
    py2 <- dpgnz(y, lambda = 15, alpha = 1/30)
    py3 <- dpgnz(y, lambda = 15, alpha = -1/30)
    plot(py0 ~ y, type = "h",
         ylim = c(0, max(c(py0, py2, py3))),
         ylab = expression(Pr(Y == y)))
    points(y + 0.1, py1, type = "h", col = 2)
    points(y - 0.3, py2, type = "h", col = 3)
    points(y + 0.3, py3, type = "h", col = 4)
    legend("topleft", bty = "n",
           col = c(1:4), lty = 1,
           legend = expression(
               Poisson(lambda == 15),
               PG(lambda == 15, alpha == 0),
               PG(lambda == 15, alpha == 1/30),
               PG(lambda == 15, alpha == -1/30)))
})

}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}
}
