% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcnt.R
\name{gcnt}
\alias{gcnt}
\title{Ajuste do Modelo Gamma Count}
\usage{
gcnt(formula, data, start = NULL, ...)
}
\arguments{
\item{formula}{Um objeto da classe \code{\link{formula}}. Se
necessária a inclusão de \emph{offset} deve-se indicá-lo como
\code{\link{offset}}.}

\item{data}{Um objeto de classe \code{data.frame} que contém as
variáveis descritas na \code{formula}.}

\item{start}{Um vetor com os valores iniciais para os parâmetros do
modelo necessários para o início do procedimento de estimação. Se
\code{NULL} as estimativas de um modelo log-linear Poisson, com
\eqn{\alpha = 0}, são utilizadas como valores iniciais, pois uma
chamada da \code{\link[stats]{glm.fit}} é feita internamente para
obtê-los. O parâmetro \eqn{\alpha} deve ser o primeiro elemento
do vetor. Os restantes devem estar na posição correspondente às
colunas da matriz gerada pelo argumento \code{formula}.}

\item{...}{Argumentos opcionais do framework de maximização numérica
\code{\link[bbmle]{mle2}}.}
}
\value{
Um objeto de classe \code{mle2}, retornado da função de
    \code{\link[bbmle]{mle2}}, usada para estimação por máxima
    verossimilhança de modelos.
}
\description{
Estima os parâmetros de um modelo Gamma Count pela
    otimização da função de log-verossimilhança definida em
    \code{\link{llgcnt}}. A sintaxe assemelha-se com a função
    \code{\link[stats]{glm}} (Generalized Linear Models).
}
\examples{

library(bbmle)

str(soja)
soja <- soja[-74, ]

m0 <- gcnt(nvag ~ bloc + umid * factor(K), data = soja)
m1 <- gcnt(nvag ~ bloc + umid + factor(K), data = soja)

anova(m0, m1)
summary(m1)

plot(profile(m1, which = "alpha"))
abline(v = 0, lty = 2, col = 2)

str(capdesfo)

n0 <- gcnt(ncap ~ est * (des + I(des^2)), data = capdesfo)
n1 <- gcnt(ncap ~ est + (des + I(des^2)), data = capdesfo)

anova(n0, n1)
summary(n0)

plot(profile(n0, which = "alpha"))

}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}.
}
