% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp.R
\name{llcmp}
\alias{llcmp}
\title{Log-Verossimilhança do Modelo Conway-Maxwell-Poisson}
\usage{
llcmp(params, y, X, offset = NULL, sumto = ceiling(max(y)^1.2))
}
\arguments{
\item{params}{Um vetor de parâmetros do modelo COM-Poisson. É
necessário que seja informado como primeiro elemento do vetor, o
valor de \eqn{\phi}. Os demais elementos são os \eqn{\beta}'s do
preditor linear \eqn{\lambda}.}

\item{y}{Um vetor de contagens, considerado como variável resposta.}

\item{X}{A matriz de delineamento do modelo.}

\item{offset}{Um vetor de valores a serem adicionados ao preditor
linear.}

\item{sumto}{Número de incrementos a serem considerados para a soma
da constante normalizadora. Como padrão, o número de incrementos
é o valor inteiro de \eqn{(\max y)^1.2}.}
}
\value{
O valor da log-verossimilhança do modelo
    Conway-Maxwell-Poisson com os parâmetros e dados informados.
}
\description{
Calcula a log-verossimilhança de um modelo COM-Poisson
    considerando os dados e as estimativas dos parâmetros informadas.
}
\details{
A função de log-verossimilhança toma a forma: \deqn{-Z - y *
    \lambda - \nu \log{y!}}, onde \eqn{Z = \sum
    \frac{\lambda^i}{(i!)^\nu}} e \eqn{\nu = \exp{\phi}}.
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}
\seealso{
\code{\link[bbmle]{mle2}}
}

