% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgnz.R
\name{llpgnz}
\alias{llpgnz}
\title{Log-Verossimilhan\enc{ç}{c}a do Modelo Poisson Generalizada}
\usage{
llpgnz(params, y, X, offset = NULL)
}
\arguments{
\item{params}{Um vetor de (estimativas dos) parâmetros do modelo de
regressão. O primeiro elemento desse vetor é o parâmetro de
dispersão do modelo e os restantes são parâmetros de
locação. Essa função retorna o negativo da log-verossimilhança
pois foi construída para ser usada na \code{\link[bbmle]{mle2}}.}

\item{y}{Um vetor com variável dependente do modelo, resposta do tipo
contagem.}

\item{X}{A matriz de delineamento correspondente ao modelo linear
ligado à média pela função de ligação log. A matriz do modelo
pode ser construída com a função
\code{\link[stats]{model.matrix}}.}

\item{offset}{Um vetor, de mesmo comprimento de \code{y}, com valores
que correspondem aos offsets (ou exposição) para cada valor
observado. Se \code{NULL}, é usado 1 como offset.}
}
\value{
O negativo da log-verossimilhança do modelo Poisson
    Generalizado para os parâmetros e dados informados.
}
\description{
Calcula a log-verossimilhança de um modelo de regressão
    para a média com distribuição Poisson Generalizada para a
    resposta de contagem (y).
}
\details{
A função de log-verossimilhança da Poisson Generalizada, na
    parametrização de modelo de regressão é:

\deqn{\ell(\lambda,\alpha,y) =
  y (\log(\lambda) - \log(1 + \alpha\lambda)) +
  (y - 1) \log(1 + \alpha y) -
  \lambda\left(\frac{1 + \alpha y}{1 + \alpha\lambda}\right) -
  \log(y!), }

em que \eqn{\alpha} é o parâmetro de dispersão e \eqn{\lambda > 0} é
    a média \eqn{E(Y) = \lambda} e \eqn{y = 0,1,ldots} é vetor
    observado da variável de contagem. Nessa parametrização,
    \eqn{V(Y) = \lambda (1 + \alpha\lambda)^2}. A Poisson
    Generalizada em a Poisson como caso particular quando \eqn{\alpha
    = 0}.

Para o modelo de regressão, um preditor linear é ligado à média pela
    função de ligação log, \eqn{\log(\lambda) = X\beta}, tal como é
    para o modelo Poisson com link log.

O espaço paramétrico de \eqn{\alpha} não limitado para o lado direito
    do zero (\eqn{\alpha} positivo) mas para o lado esquerdo
    (\eqn{\alpha} negativo) o limite inferior é dependente do
    parâmetro \eqn{\lambda} e dos valores observados de
    \eqn{y}. Valores não finitos podem ser retornados durante a
    estimação quando \eqn{1 + \alpha\lambda} ou \eqn{1 + \alpha y}
    não forem maiores que zero.
}
\examples{

set.seed(123)
y <- rpois(10, lambda = 5)

# Log-verossimilhança pela Poisson.
sum(dpois(y, lambda = 5, log = TRUE))

# Log-verossimilhança pela PGNZ usando alpha = 0
llpgnz(params = c(0, log(5)), y = y, X = cbind(y * 0 + 1))

set.seed(121)
y <- rpois(100, lambda = exp(1))
X <- cbind(0 * y + 1)

grid <- expand.grid(alpha = seq(-0.1, 0.4, by = 0.01),
                    lambda = seq(0.1, 2.1, by = 0.025))

grid$ll <- apply(grid, MARGIN = 1,
                 FUN = function(vec) {
                     llpgnz(params = vec, y = y, X = X, offset = NULL)
                 })

library(latticeExtra)

levelplot(ll ~ alpha + lambda, data = grid) +
    layer(panel.abline(v = 0, h = 1, lty = 2))

}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}
}
\seealso{
\code{\link[bbmle]{mle2}}.
}

