% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRDCr.R
\docType{package}
\name{MRDCr}
\alias{MRDCr}
\alias{MRDCr-package}
\title{Modelos de Regress\enc{ã}{a}o para Dados de Contagem}
\description{
Esse é um pacote com o material do Curso "Modelos de
    Regressão para Dados de Contagem" oferecido pelos autores na 61
    RBRAS (\url{rbras2016.ufba.br}) que aconteceu entre 23 e 25 de
    Maio em Salvador, Bahia.

    O objetivo desse Curso é apresentar diferentes modelos de
    regressão para análise de dados de contagem, discutir suas
    principais propriedades e ilustrar suas aplicações em dados reais
    por meio do software R. Dentre os modelos considerados estão o
    modelo Poisson, algumas extensões para acomodação de super (ou
    sub) dispersão (Quasi-Poisson, COM-Poisson, Count-Gamma,
    Binomial-Nagativo), modelos para dados inflacionados de zeros e
    com a inclusão de efeitos aleatórios.

    Neste pacote estão disponíveis as vignettes que mostram como
    coordenar funções do R para análise dos dados, os slides do
    Curso, as funções e alguns dos dados utitizados nos estudos de
    caso.
}

