% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.beeswarm.R
\name{panel.beeswarm}
\alias{panel.beeswarm}
\title{Função para Gráfico de Dispersão BeesWarm com a Lattice}
\usage{
panel.beeswarm(x, y, subscripts, r, ...)
}
\arguments{
\item{x, y, subscripts, ...}{Argumentos passados para a
função \code{\link[lattice]{xyplot}}.}

\item{r}{Valor para espaçamento entre os pontos de mesmas
coordenadas.}
}
\value{
Painel \code{xyplot} padrão, com as coordenadas \code{x}
    devidamente corrigidas para agrupamento lado a lado.
}
\description{
Essa função permite a construção de gráficos de
    dispersão unidimensionais similares a função
    \code{\link{stripchart}}. Porém com pontos de mesmas coordenadas
    exibidos lado a lado e não sobrepostos.
}
\examples{
library(lattice)

set.seed(2016)
da <- data.frame(x = rep(letters[1:5], 10), y = rpois(50, 5))

xyplot(y ~ x, data = da, jitter.x = TRUE)
xyplot(y ~ x, data = da, panel = panel.beeswarm, r = 0.1)
xyplot(y ~ x, data = da, panel = panel.beeswarm, r = 0.05)

xyplot(drat ~ carb | am, data = mtcars, jitter.x = TRUE)
xyplot(drat ~ carb | am, data = mtcars, panel = panel.beeswarm, r = 0.2)

}
\author{
Walmes Zeviani
}
\references{
Aron Eklund (2016). beeswarm: The Bee Swarm Plot, an
    Alternative to Stripchart. R package version
    0.2.3. \url{https://CRAN.R-project.org/package=beeswarm}
}

