% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.groups.segplot.R
\name{panel.groups.segplot}
\alias{panel.groups.segplot}
\title{Painel para Fazer Intervalos para Grupos no \code{segplot()}}
\usage{
panel.groups.segplot(x, y, z, centers, groups, gap = NULL, data, subscripts,
  ...)
}
\arguments{
\item{x, y, z, centers, data, subscripts, ...}{Argumentos passados para a
\code{\link[latticeExtra]{segplot}}.}

\item{groups}{A variável (\code{factor}) de agrupamento, de mesmo
comprimento de \code{lwr} e \code{upr}.}

\item{gap}{Escalar que representa a distância entre os segmentos. O
valor default é 0,1. Como um fator com \eqn{k} níveis é
representado pelos números inteiros \eqn{1, 2, \ldots, k}, então
\eqn{0 \leq \textrm{gap} < 1/k}. Se for usado um valor negativo,
os intervalos serão apresentados em ordem inversa.}
}
\value{
A função é passada para o argumento \code{panel} e retorna
    elementos necessários para a \code{\link[latticeExtra]{segplot}}.
}
\description{
Essa função permite fazer intervalos de confiança (ou
    com barras de erro) em gráficos da \code{latticeExtra::segplot()}
    para grupos (\code{groups =}) de tal forma que eles não fiquem
    sobrepostos.
}
\examples{

library(latticeExtra)

m0 <- lm(log(breaks) ~ wool * tension, data = warpbreaks)

pred <- with(warpbreaks, expand.grid(KEEP.OUT.ATTRS = TRUE,
                                     wool = levels(wool),
                                     tension = levels(tension)))

pred <- cbind(pred,
              predict(m0, newdata = pred, interval = "confidence"))
str(pred)

segplot(wool ~ lwr + upr, centers = fit, data = pred,
        draw = FALSE, horizontal = FALSE)

segplot(wool ~ lwr + upr, centers = fit, data = pred,
        draw = FALSE, horizontal = FALSE,
        groups = tension, gap = NULL,
        panel = panel.groups.segplot)

}
\author{
Walmes Zeviani
}
\seealso{
\code{\link[latticeExtra]{segplot}}.
}

