% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soja.R
\name{soja}
\alias{soja}
\title{Umidade do Solo e Doses de Potássio na Cultura da Soja}
\format{Um \code{data.frame} com 75 observações e 5 variáveis, em que

\describe{

\item{\code{K}}{Fator de níveis métricos. Dose de potássio aplicado
    ao solo na forma de adubação, em mg dm\eqn{^{-3}}.}

\item{\code{umid}}{Fator de níveis categóricos. Faixas de umidade do
    solo. Procurou-se por meio de pesagens diárias dos vasos, manter
    a umidade do solo nas faixas de 35 a 40; 47,5 a 52,5; e 60 a 65
    do volume total de poros.}

\item{\code{bloc}}{Fator categórico para controle local no
    experimento. Foram controladas as variações de ambiente da casa
    de vegetação, como exposição ao sol, e mão de obra ao longo do
    experimento,  como irrigação e capina dos vasos e controle de
    doenças.}

\item{\code{ngra}}{Número de grãos por vaso na colheita da soja.}

\item{\code{nvag}}{Número de vagens viáveis por vaso na colheita da
    soja.}

}}
\source{
Serafim, Milson Evaldo; Ono, Fábio Benedito; Zeviani, Walmes
    Marques; Novelino, José Oscar; Silva, Joil
    Vilhalva. (2012). Umidade do solo e doses de potássio na cultura
    da soja. Revista Ciência Agronômica, 43(2),
    222-227. https://dx.doi.org/10.1590/S1806-66902012000200003,
    \url{http://www.scielo.br/pdf/rca/v43n2/a03v43n2.pdf}
}
\description{
Variáveis de contagem de um experimento fatorial 5
    \eqn{\times} 3 que estudou níveis de adubação com potássio e
    níveis de umidade do solo nos componentes de produção da soja. O
    experimento foi instalado em casa de vegetação no delineamento de
    blocos casualizados completos e a unidade experimental foi um
    vaso com duas plantas de soja. No experimento foram medidas
    várias respostas, sendo que o número de vagens por vaso e o
    número de grãos por vaso foram as variáveis de contagem.
}
\details{
A observação número 74 (bloco V, K 120, umid 62,5) foi uma
    parcela onde as plantas não tiveram um desenvolvimento
    normal. Portanto, pode-se remover esta observação para análise
    dos dados.
}
\examples{

library(lattice)

data(soja)
str(soja)

xtabs(~umid + K, data = soja[-75, ])

xyplot(nvag + ngra ~ K, groups = umid, outer = TRUE,
       data = soja[-74, ],
       type = c("p", "a"), scales = "free",
       ylab = NULL,
       xlab = expression("Nível de potássio aplicado"~(mg~dm^{-3})),
       auto.key = list(title = "Umidade do solo (\%)",
                       cex.title = 1,
                       columns = 3),
       strip = strip.custom(
           factor.levels = c("Número de vagens",
                             "Número de grãos")))

}

