% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp.R
\name{convergencez}
\alias{convergencez}
\title{Avalia\enc{çã}{ca}o da Converg\enc{ê}{e}ncia da Constante Normalizadora}
\usage{
convergencez(model, tol = 1e-04, incremento = 10, maxit = 150,
  plot = TRUE)
}
\arguments{
\item{model}{Objeto resultante da fun\enc{çã}{ca}o \code{\link[MRDCr]{cmp}}.}

\item{tol}{Crit\enc{é}{e}rio de parada do algoritmo, representa o valor
tolerado para a diferença de \eqn{\frac{\lambda^i}{(i!)^\nu} -
0}, pois no limite \eqn{i \rightarrow \infty} o incremente
\eqn{\frac{\lambda^i}{(i!)^\nu}} tende a 0.}

\item{incremento}{Número de incrementos da soma a serem considerados
a cada iteração. Padrão definido como 10, ou seja, a cada
iteração 10 incrementos são calculados.}

\item{maxit}{Número máximo de iterações a serem realizadas pelo
algoritmo. Se este número for atingido e o critério de tolerância
não for atendido, uma mensagem de aviso será exibida.}

\item{plot}{Argumento lógico. Se \code{TRUE} (padrão) os gráficos dos
incrementos daa constantes, calculadas para cada observação são
exibidos.}
}
\value{
Uma lista com os incrementos das constantes Z,
    \eqn{Z(\lambda, \phi)} da distribuição COM-Poisson, calculados
    para cada observação.
}
\description{
Avalia a converg\enc{ê}{e}ncia da constante de normaliza\enc{çã}{ca}o de um
    modelo COM-Poisson definida por: \deqn{Z = \sum
    \frac{\lambda^i}{(i!)^\nu}}, em que o parâmetro \eqn{\nu} \enc{é}{e}
    tomado como \eqn{\exp{\phi}}.
}
\examples{

m1 <- cmp(ncap ~ est * (des + I(des^2)), data = capdesfo)
tablez <- convergencez(m1)
str(tablez)

m2 <- cmp(ncap ~ dexp + I(dexp^2), data = capmosca)
tablez <- convergencez(m2)
str(tablez)

}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}

