% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp.R
\name{dcmp}
\alias{dcmp}
\title{Probabilidades do Modelo Conway-Maxwell-Poisson}
\usage{
dcmp(y, loglambda, phi, sumto)
}
\arguments{
\item{y}{Valor da contagem}

\item{loglambda}{Logaritmo neperiano do parâmetro \eqn{\lambda} da
distribuição}

\item{phi}{Valor do parâmetro \eqn{\phi}, definido como
\eqn{\log(\nu)} na distribuição de probabilidades COM-Poisson}

\item{sumto}{Número de incrementos a serem considerados para a soma
da constante normalizadora Z.}
}
\description{
Calcula as probabilidades conforme modelo COM-Poisson
}
\examples{
dpois(5, 5)
dcmp(5, log(5), 0, sumto = 20)

x <- 0:30
px1 <- dpois(x, 15)
px2 <- dcmp(x, log(915), log(2.5), sumto = 50)
px3 <- dcmp(x, log(2.2), log(0.3), sumto = 50)

plot(y = px2, x = x, type = "h", lwd = 2)
lines(y = px1, x = x - 0.2, type = "h", lwd = 2, col = 4)
lines(y = px3, x = x + 0.2, type = "h", lwd = 2, col = 2)
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}

